Changed Progress to bar + float Added Scene to SimSystem (badly)

+++ include/FluidSim/DistanceField.h
@@ -15,7 +15,7 @@ private:
 
 public:
     DistanceField(glm::vec2 pos, int width, int height, unsigned char* data, std::string& file);
-    DistanceField(glm::vec2 pos, int width, int height, unsigned char* data, std::string& file, int* progress, std::string* progress_msg);
+    DistanceField(glm::vec2 pos, int width, int height, unsigned char* data, std::string& file, float* progress, std::string* progress_msg);
     ~DistanceField();
 
     int distance(glm::vec2 pos);
@@ -33,7 +33,7 @@ private:
 
     std::vector<Field*> _grid;
 
-    int* _progress{};
+    float* _progress{};
     std::string* _progress_msg{};
     std::string _file_name;
